NF: make PartialSearch its own class
Seems big enough to be worth it
