+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -941,78 +941,6 @@ public class CollectionTask<Progress, Result> extends BaseAsyncTask<Void, Progre
         }
     }
 
-    /**
-     * A class allowing to send partial search result to the browser to display while the search ends
-     */
-    public static class PartialSearch implements ProgressSenderAndCancelListener<List<Long>> {
-        private final List<CardBrowser.CardCache> mCards;
-        private final int mColumn1Index, mColumn2Index;
-        private final int mNumCardsToRender;
-        private final ProgressSenderAndCancelListener<List<CardBrowser.CardCache>> mCollectionTask;
-        private final Collection mCol;
-
-        public PartialSearch(List<CardBrowser.CardCache> cards, int columnIndex1, int columnIndex2, int numCardsToRender, ProgressSenderAndCancelListener<List<CardBrowser.CardCache>> collectionTask, Collection col) {
-            mCards = new ArrayList<>(cards);
-            mColumn1Index = columnIndex1;
-            mColumn2Index = columnIndex2;
-            mNumCardsToRender = numCardsToRender;
-            mCollectionTask = collectionTask;
-            mCol = col;
-        }
-
-        @Override
-        public boolean isCancelled() {
-            return mCollectionTask.isCancelled();
-        }
-
-
-        /**
-         * @param cards Card ids to display in the browser. It is assumed that it is as least as long as mCards, and that
-         *             mCards[i].cid = cards[i].  It add the cards in cards after `mPosition` to mCards
-         */
-        public void add(@NonNull List<Long> cards) {
-            while (mCards.size() < cards.size()) {
-                mCards.add(new CardBrowser.CardCache(cards.get(mCards.size()), mCol, mCards.size()));
-            }
-        }
-
-
-        @Override
-        public void doProgress(@NonNull List<Long> value) {
-            // PERF: This is currently called on the background thread and blocks further execution of the search
-            // PERF: This performs an individual query to load each note
-            add(value);
-            for (CardBrowser.CardCache card : mCards) {
-                if (isCancelled()) {
-                    Timber.d("doInBackgroundSearchCards was cancelled so return");
-                    return;
-                }
-                card.load(false, mColumn1Index, mColumn2Index);
-            }
-            mCollectionTask.doProgress(mCards);
-        }
-
-
-        public ProgressSender<Long> getProgressSender() {
-            return new ProgressSender<Long>() {
-                private final List<Long> mRes = new ArrayList<>();
-                private boolean mSendProgress = true;
-                @Override
-                public void doProgress(@Nullable Long value) {
-                    if (!mSendProgress) {
-                        return;
-                    }
-                    mRes.add(value);
-                    if (mRes.size() >= mNumCardsToRender) {
-                        PartialSearch.this.doProgress(mRes);
-                        mSendProgress = false;
-                    }
-                }
-            };
-        }
-    }
-
-
     public static class RebuildCram implements TaskDelegate<Void, StudyOptionsFragment.DeckStudyData> {
         public StudyOptionsFragment.DeckStudyData task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<Void> collectionTask) {
             Timber.d("doInBackgroundRebuildCram");
